# app/schemas/3dbagapi_spec.yaml
---
openapi: 3.0.3
info:
  title: '3DBAG API'
  description: |-
    The [3DBAG](https://3dbag.nl) is an up-to-date dataset containing 3D building models of the Netherlands. 
    The 3DBAG is open data. It contains 3D models at multiple levels of detail, 
    which are generated by combining two open data sets: 
    the building data from the BAG and the height data from the AHN. 
    The 3DBAG is updated regularly, keeping it up-to-date with the latest openly 
    available building stock and elevation information.
    
    The 3DBAG is maintained by 3D GeoInformation group and 3DGI.
    
    Copyright (c) 2023 3D GeoInformation group + 3DGI.
  version: 0.1.0
  contact:
    name: "3DBAG support"
    email: info@3dbag.nl
    url: https://docs.3dbag.nl/en/contact/
  license:
    name: CC-BY-4.0
    url: https://creativecommons.org/licenses/by/4.0
  x-logo:
    url: https://3dgi.nl/img/Logo-02.png
servers:
  - url: api.3dbag.nl
    description: 3DBAG API production server
tags:
  - name: Capabilities
    description: |-
      Essential characteristics of this API
  - name: Data
    description: |-
      Access to data (features)
paths:
  '/':
    get:
      tags:
        - Capabilities
      summary: Landing Page
      description: |-
        The landing page provides links to the API definition, the conformance
        statements and to the feature collections in this dataset.
      operationId: getLandingPage
      responses:
        '200':
          $ref: '#/components/responses/LandingPage'
        '500':
          $ref: '#/components/responses/ServerError'
  '/conformance':
    get:
      tags:
        - Capabilities
      summary: Information about specifications that this API conforms to
      description: |-
        A list of all conformance classes specified in a standard that the
        server conforms to.
      operationId: getConformanceDeclaration
      responses:
        '200':
          $ref: '#/components/responses/ConformanceDeclaration'
        '500':
          $ref: '#/components/responses/ServerError'
  '/collections':
    get:
      tags:
        - Capabilities
      summary: The feature collections in the dataset
      operationId: getCollections
      responses:
        '200':
          $ref: '#/components/responses/Collections'
        '500':
          $ref: '#/components/responses/ServerError'
  '/collections/pand':
    get:
      tags:
        - Capabilities
      summary: 3D building models based on the 'pand' layer of the BAG data set
      operationId: describeCollection
      responses:
        '200':
          $ref: '#/components/responses/Collection'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  '/collections/pand/items':
    get:
      tags:
        - Data
      summary: fetch pand features
      description: |-
        Fetch features of the feature collection with id `pand`.
      operationId: getFeatures
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/bbox'
        - $ref: '#/components/parameters/crs'
        - $ref: '#/components/parameters/bbox-crs'
      responses:
        '200':
          $ref: '#/components/responses/Features'
        '400':
          $ref: '#/components/responses/InvalidParameter'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  '/collections/pand/items/{featureId}':
    get:
      tags:
        - Data
      summary: fetch a single 3DBAG feature
      description: |-
        Fetches a 3DBAG object/feature in CityJSONFeatures format.
        The feature includes all Levels of Detail.
      operationId: getFeatureId
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/featureId'
        - $ref: '#/components/parameters/crs'
      responses:
        '200':
          $ref: '#/components/responses/Feature'
        '400':
          $ref: '#/components/responses/InvalidParameter'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  parameters: 
    bbox:
      name: bbox
      in: query
      description: |-
        Only features that have a geometry that intersects the bounding box are selected.
        The bounding box is provided as four numbers:

        * Lower left corner, coordinate axis 1 (min. x)
        * Lower left corner, coordinate axis 2 (min. y)
        * Upper right corner, coordinate axis 1 (max. x)
        * Upper right corner, coordinate axis 2 (max. y)

        The default coordinate reference system is Amersfoort / RD New + NAP height, EPSG:7415
        (https://www.opengis.net/def/crs/EPSG/0/7415).
        Currently only `https://www.opengis.net/def/crs/EPSG/0/7415` is supported as `bbox-crs`.
      required: false
      schema:
        type: array
        minItems: 4
        maxItems: 6
        items:
          type: number
      style: form
      explode: false
    bbox-crs:
      name: bbox-crs
      description: |-
        Asserts the CRS used for the coordinate values of the bbox parameter. 
        The default is EPSG:7415.
      example: "https://www.opengis.net/def/crs/EPSG/0/7415"
      in: query
      required: false
      schema:
        type: string
        format: uri
      style: form
      explode: false
    crs:
      name: crs
      description: |-
        Required CRS for the response. 
        The default is EPSG:7415.
      example: "https://www.opengis.net/def/crs/EPSG/0/7415"
      in: query
      required: false
      schema:
        type: string
        format: uri
      style: form
      explode: false
    featureId:
      name: featureId
      in: path
      description: 3DBAG [`identificatie`](https://docs.3dbag.nl/en/schema/attributes/#identificatie)
      example: "NL.IMBAG.Pand.0851100000000564-0"
      required: true
      schema:
        type: string
    limit:
      name: limit
      in: query
      description: |-
        The optional limit parameter limits the number of items that are presented in the response document.

        Only items are counted that are on the first level of the collection in the response document.
        Nested objects contained within the explicitly requested items shall not be counted.

        Minimum = 1. Maximum = 10000. Default = 10.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 10000
        default: 10
      style: form
      explode: false
    offset:
      name: offset
      in: query
      description: |-
        Offset parameter used for pagination.
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      style: form
      explode: false
  schemas: 
    collection:
      type: object
      required:
        - id
        - links
      properties:
        id:
          description: identifier of the collection used, for example, in URIs
          type: string
          example: "id"
        title:
          description: human readable title of the collection
          type: string
          example: "title"
        description:
          description: a description of the features in the collection
          type: string
          example: "A description"
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
          example:
            - href: http://data.example.com/buildings
              rel: item
            - href: http://example.com/concepts/buildings.html
              rel: describedby
              type: text/html
        extent:
          $ref: '#/components/schemas/extent'
        itemType:
          description: indicator about the type of the items in the collection (the default value is 'feature').
          type: string
          default: feature
        crs:
          description: the list of coordinate reference systems supported by the service
          type: array
          items:
            type: string
          default:
            - https://www.opengis.net/def/crs/EPSG/0/7415
          example:
            - https://www.opengis.net/def/crs/EPSG/0/7415
    collections:
      type: object
      required:
        - links
        - collections
      properties:
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
        collections:
          type: array
          items:
            $ref: '#/components/schemas/collection'
    confClasses:
      type: object
      required:
        - conformsTo
      properties:
        conformsTo:
          type: array
          items:
            type: string
            example: "https://cityjson.org/specs/1.1.1/"
    extent:
      description: |-
        The extent of the features in the collection. In the Core only spatial and temporal
        extents are specified. Extensions may add additional members to represent other
        extents, for example, thermal or pressure ranges.
  
        An array of extents is provided for each extent type (spatial, temporal). The first item
        in the array describes the overall extent of the data. All subsequent items describe more
        precise extents, e.g., to identify clusters of data. Clients only interested in the
        overall extent will only need to access the first extent in the array.
      type: object
      properties:
        spatial:
          description: |-
            The spatial extent of the features in the collection.
          type: object
          properties:
            bbox:
              description: |-
                One or more bounding boxes that describe the spatial extent of the dataset.
                In the Core only a single bounding box is supported.
  
                Extensions may support additional areas.
                The first bounding box describes the overall spatial
                extent of the data. All subsequent bounding boxes describe
                more precise bounding boxes, e.g., to identify clusters of data.
                Clients only interested in the overall spatial extent will
                only need to access the first bounding box in the array.
              type: array
              minItems: 1
              items:
                description: |-
                  Each bounding box is provided as four or six numbers, depending on
                  whether the coordinate reference system includes a vertical axis
                  (height or depth):
  
                  * Lower left corner, coordinate axis 1
                  * Lower left corner, coordinate axis 2
                  * Minimum value, coordinate axis 3 (optional)
                  * Upper right corner, coordinate axis 1
                  * Upper right corner, coordinate axis 2
                  * Maximum value, coordinate axis 3 (optional)
  
                  The default coordinate reference system is Amersfoort / RD New + NAP height, EPSG:7415
                  (https://www.opengis.net/def/crs/EPSG/0/7415).
                  Currently only `https://www.opengis.net/def/crs/EPSG/0/7415` is supported as `crs`.
    
                  If a feature has multiple spatial geometry properties, it is the decision of the
                  server whether only a single spatial geometry property is used to determine
                  the extent or all relevant geometries.
                type: array
                minItems: 4
                maxItems: 6
                items:
                  type: number
                example:
                  - -180
                  - -90
                  - 180
                  - 90
            crs:
              description: |-
                Coordinate reference system of the returned coordinates. 
                Currently only `https://www.opengis.net/def/crs/EPSG/0/7415`
                is supported.
              type: string
              enum:
                - 'https://www.opengis.net/def/crs/EPSG/0/7415'
              default: 'https://www.opengis.net/def/crs/EPSG/0/7415'
    exception:
      type: object
      required:
        - code
      properties:
        code:
          type: string
        description:
          type: string
        name:
          type: string
    featureCityJSON:
      type: object
      properties:
        feature:
          type: object
        id:
          type: string
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
    featureCollection:
      type: object
      required:
        - type
        - features
      properties:
        type:
          type: string
          enum:
            - FeatureCollection
        features:
          type: array
          items:
            $ref: '#/components/schemas/featureCityJSON'
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
        timeStamp:
          type: string
          format: date-time
        numberMatched:
          type: integer
          minimum: 0
        numberReturned:
          type: integer
          minimum: 0
    landingPage:
      type: object
      required:
        - links
      properties:
        title:
          type: string
          example: Buildings
        description:
          type: string
          example: Access to data about buildings in the Netherlands.
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
    link:
      type: object
      required:
        - href
        - rel
      properties:
        href:
          type: string
          example: https://api.3dbag.nl/collections/pand/items/NL.IMBAG.Pand.1655100000500568
        rel:
          type: string
          example: alternate
        type:
          type: string
          example: application/city+json
        hreflang:
          type: string
          example: en
        title:
          type: string
          example: Spui 70, 2511 BT Den Haag
        length:
          type: integer
  responses:
    LandingPage:
      description: |-
        The landing page provides links to the API definition 
        (link relations `service-desc` and `service-doc`), 
        the Conformance declaration (path `/conformance`, 
        link relation `conformance`), and the Feature 
        Collections (path `/collections`, link relation 
        `data`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/landingPage'
          example:
            title: 3DBAG API
            description: API for 3DBAG buildings of LoD 1.2, 1.3 & 2.2.
            links:
              - href: https://api.3dbag.nl
                rel: self
                type: application/json
                title: this document
              - href: https://api.3dbag.nl/api
                rel: service-desc
                type: application/vnd.oai.openapi+json;version=3.0
                title: the API definition
              - href: https://api.3dbag.nl/api.html
                rel: service-doc
                type: text/html
                title: the API documentation
              - href: https://api.3dbag.nl/conformance
                rel: conformance
                type: application/json
                title: Conformance classes implemented by this server
              - href: https://api.3dbag.nl/collections
                rel: data
                type: application/json
                title: Information about the feature collections
    ServerError:
      description: |-
        A server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/exception'
    NotFound:
      description: |-
        The requested URI was not found. Also raised when the requested `featureId` was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/exception'
    InvalidParameter:
      description: |-
        A query parameter has an invalid value.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/exception'
    ConformanceDeclaration:
      description: |-
        The URIs of all conformance classes supported by the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/confClasses'
          example:
            conformsTo:
              - 'https://cityjson.org/specs/1.1.1/'
    Collections:
      description: |-
        The feature collections shared by this API.
  
        The dataset is organized as one or more feature collections. This resource
        provides information about and access to the collections.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/collections'
          example:
            links:
              - href: https://api.3dbag.nl/collections.json
                rel: self
                type: application/json
                title: this document
            collections:
              - id: pand
                title: 3D pand
                description: The 3DBAG pand layer
                extent:
                  spatial:
                    bbox:
                      - - 10000
                        - 306250
                        - 287760
                        - 623690
                    crs: https://www.opengis.net/def/crs/EPSG/0/7415
                  temporal:
                    interval:
                      - null
                      - '2019-12-31T24:59:59Z' # TODO: '2019' is the max bouwjaar in the BAG. I just added the rest, because a timestamp is required. Not sure if we should reference the BAG dates or the AHN dates.
                itemType: feature
                crs:
                  - https://www.opengis.net/def/crs/EPSG/0/7415
                links:
                  - href: https://api.3dbag.nl/collections/pand
                    rel: items
                    type: application/city+json
                    title: Pand
                  # TODO: add license
                  - href: https://creativecommons.org/licenses/by/4.0/
                    rel: license
                    type: text/html
                    title: CC BY 4.0
                  - href: https://creativecommons.org/licenses/by/4.0/rdf
                    rel: license
                    type: application/rdf+xml
                    title: CC BY 4.0
    Collection:
      description: |-
        Information about the feature collection with id `collectionId`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/collection'
          example:
            id: pand
            title: 3D pand # TODO: give title
            description: The 3DBAG pand layer # TODO: describe
            extent:
              spatial:
                bbox:
                  - - 10000
                    - 306250
                    - 287760
                    - 623690
                crs: https://www.opengis.net/def/crs/EPSG/0/7415
              temporal:
                interval:
                  - null
                  - '2019-12-31T24:59:59Z' # TODO: '2019' is the max bouwjaar in the BAG. I just added the rest, because a timestamp is required. Not sure if we should reference the BAG dates or the AHN dates.
            itemType: feature
            crs:
              - https://www.opengis.net/def/crs/EPSG/0/7415
            links:
              - href: https://api.3dbag.nl/collections/pand
                rel: items
                type: application/city+json
                title: Pand
              # TODO: add license
              - href: https://creativecommons.org/licenses/by/4.0/
                rel: license
                type: text/html
                title: CC BY 4.0
              - href: https://creativecommons.org/licenses/by/4.0/rdf
                rel: license
                type: application/rdf+xml
                title: CC BY 4.0
    Features:
      description: |-
        The response is a document consisting of features in the collection.
        The features included in the response are determined by the server
        based on the query parameters of the request. To support access to
        larger collections without overloading the client, the API supports
        paged access with links to the next page, if more features are selected
        that the page size.
        
        The `bbox` parameter can be used to select only a
        subset of the features in the collection (the features that are in the
        bounding box). The `bbox` parameter matches all features
        in the collection that are not associated with a location, too.

        The `limit` parameter may be used to control the subset of the
        selected features that should be returned in the response, the page size.
        Each page includes information about the number of selected and
        returned features (`numberMatched` and `numberReturned`) as well as
        links to support paging (link relation `next`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/featureCollection'
    Feature:
      description: |-
        Fetch a single feature with id `featureId` in *CityJSONFeature* format.
        
        Feature hierarchy is represented by by the link relations `child` and `parent`.
      content:
        application/city+json:
          schema:
            $ref: '#/components/schemas/featureCityJSON'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY